fix unconvert linting
    bool_test.go:29:22: unnecessary conversion (unconvert)
        return triStateValue(*v)
                            ^
    count.go:16:18: unnecessary conversion (unconvert)
            *i = countValue(*i + 1)
                           ^
    uint64.go:30:15: unnecessary conversion (unconvert)
        return uint64(v), nil
                     ^

+++ uint64.go
@@ -27,7 +27,7 @@ func uint64Conv(sval string) (interface{}, error) {
 	if err != nil {
 		return 0, err
 	}
-	return uint64(v), nil
+	return v, nil
 }
 
 // GetUint64 return the uint64 value of a flag with the given name
